Extracts `useScrollDirection` to a separated file
